after jumping, set current player space to fix bug
After jumping, clicking on an execute button would advance the player
to the next space from where they were before the jump, which would place
them in the wrong place on the board.  Now, jumping updates the player so
they are moved to the board place where they jumped, so everything is synced.